Update Main
Strangely it only likes the trig conversion functions if called within the main program. I think this may be do to optimization of code by the compiler. In doing so it tosses "unused values" out. This can revert values back to default settings. For testing purposes the family of 6 trig functions works. 